Community to EAP conversion -- Drop unstable annotation API
For now, we drop all of these tests, however, this situation should be
changed upstream since we should be able to run this testsuite even if
we have provisioned to a higher stability level beyond "preview".

This work is pending to be done upstream. Right now, this test module
fails when running in defaul stability level, so let's drop from here
until the situation is fixed.