Community to EAP conversion -- Ignore WildFly Model Test Controller versions
The ModelTestControllerVersion stablishes a set of legacy controller versions used to test transformers against legacy versions. Some of those Controller versions are meant onlt for upstream to test the Community - Community stability level transformation. In JBoss EAP we do not use those legacy controller version, so we need to ignore them so the tests gets ignored when these versions are in use.